Haveli-GAN/requirements.txt
# Haveli-GAN Requirements
# Compatible with Python 3.10 and CUDA 12.8

# Core deep learning framework
torch>=2.5.0
torchvision>=0.20.0
torchaudio>=2.5.0

# Image processing
Pillow>=10.0.0

# Utilities
tqdm>=4.65.0

# Data processing and artificial damage generation
opencv-python>=4.12.0

# Note: For CUDA 12.8 support, install PyTorch using:
# conda install pytorch torchvision torchaudio pytorch-cuda=12.1 -c pytorch -c nvidia

# Alternative pip installation (if conda not available):
# pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu121
